src/requirements.txt
azure-functions
azure-ai-documentintelligence
azure-core
sqlalchemy
psycopg2-binary
openai